# webui/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateResult": {
        "description": "实际建出来的可能比请求的少：重复下载与需要二次解析的会被拦掉",
        "properties": {
          "created": {
            "title": "Created",
            "type": "integer"
          },
          "requested": {
            "title": "Requested",
            "type": "integer"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskView"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "requested",
          "created",
          "tasks"
        ],
        "title": "CreateResult",
        "type": "object"
      },
      "CreateTasksRequest": {
        "properties": {
          "episodes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "maxItems": 2000,
            "minItems": 1,
            "title": "Episodes",
            "type": "array"
          },
          "options": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          }
        },
        "required": [
          "episodes"
        ],
        "title": "CreateTasksRequest",
        "type": "object"
      },
      "DeleteResult": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "integer"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "DeleteResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PauseResult": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          },
          "streams_affected": {
            "title": "Streams Affected",
            "type": "integer"
          },
          "updated": {
            "title": "Updated",
            "type": "integer"
          }
        },
        "required": [
          "updated",
          "streams_affected",
          "status"
        ],
        "title": "PauseResult",
        "type": "object"
      },
...
      "RetryResult": {
        "properties": {
          "retried": {
            "title": "Retried",
            "type": "integer"
          }
        },
        "required": [
          "retried"
        ],
        "title": "RetryResult",
        "type": "object"
      },
...
      "TaskIdsRequest": {
        "properties": {
          "task_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 2000,
            "minItems": 1,
            "title": "Task Ids",
            "type": "array"
          }
        },
        "required": [
          "task_ids"
        ],
        "title": "TaskIdsRequest",
        "type": "object"
      },
      "TaskList": {
        "properties": {
          "ascending": {
            "title": "Ascending",
            "type": "boolean"
          },
          "sort_by": {
            "title": "Sort By",
            "type": "string"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskView"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "tasks",
          "sort_by",
          "ascending"
        ],
        "title": "TaskList",
        "type": "object"
      },
      "TaskSnapshot": {
        "description": "全量快照\n\n`cursor` 是发快照那一刻的事件编号，前端带着它连 WebSocket（`?since=cursor`），\n否则快照与增量之间会漏掉一段",
        "properties": {
          "completed": {
            "items": {
              "$ref": "#/components/schemas/TaskView"
            },
            "title": "Completed",
            "type": "array"
          },
          "cursor": {
            "title": "Cursor",
            "type": "integer"
          },
          "downloading": {
            "items": {
              "$ref": "#/components/schemas/TaskView"
            },
            "title": "Downloading",
            "type": "array"
          }
        },
        "required": [
          "cursor",
          "downloading",
          "completed"
        ],
        "title": "TaskSnapshot",
        "type": "object"
      },
      "TaskView": {
        "description": "与 `web/download/view.py` 的 `task_view()` 一一对应\n\n**两边字段必须同步**：那边加了字段这里不加，前端就看不见；\n这里多写一个那边没有的，前端会拿到 undefined",
        "properties": {
          "audio_quality": {
            "default": "",
            "title": "Audio Quality",
            "type": "string"
          },
          "completed_time": {
            "default": 0,
            "title": "Completed Time",
            "type": "integer"
          },
          "cover_id": {
            "default": "",
            "title": "Cover Id",
            "type": "string"
          },
          "created_time": {
            "default": 0,
            "title": "Created Time",
            "type": "integer"
          },
          "download_path": {
            "default": "",
            "title": "Download Path",
            "type": "string"
          },
          "downloaded_size": {
            "default": 0,
            "title": "Downloaded Size",
            "type": "integer"
          },
          "duration": {
            "default": 0,
            "title": "Duration",
            "type": "integer"
          },
          "file_name": {
            "default": "",
            "title": "File Name",
            "type": "string"
          },
          "folder": {
            "default": "",
            "title": "Folder",
            "type": "string"
          },
          "info_label": {
            "default": "",
            "title": "Info Label",
            "type": "string"
          },
          "progress": {
            "default": 0,
            "title": "Progress",
            "type": "integer"
          },
          "speed": {
            "default": 0,
            "title": "Speed",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "status_label": {
            "default": "",
            "title": "Status Label",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "total_size": {
            "default": 0,
            "title": "Total Size",
            "type": "integer"
          },
          "type": {
            "items": {
              "type": "string"
            },
            "title": "Type",
            "type": "array"
          },
          "url": {
            "default": "",
            "title": "Url",
            "type": "string"
          },
          "video_codec": {
            "default": "",
            "title": "Video Codec",
            "type": "string"
          },
          "video_quality": {
            "default": "",
            "title": "Video Quality",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "title",
          "status"
        ],
        "title": "TaskView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Bili23 Downloader WebUI",
    "version": "2.15.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/tasks": {
      "get": {
        "description": "全量快照\n\n带上 `cursor`：前端拿它去连 WebSocket（`?since=cursor`），服务端会把这之后的\n增量补发出来。没有这个游标的话，快照与增量之间会漏掉一段",
        "operationId": "task_snapshot_api_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Task Snapshot",
        "tags": [
          "events"
        ]
      },
      "post": {
        "description": "创建下载任务\n\n`task_manager.create()` 内部会做**重复下载判定**与**是否需要二次解析**的判断，\n所以实际建出来的可能比传进来的少 —— 返回里如实给出两个数字，\n不然前端会以为「点了没反应」",
        "operationId": "create_tasks_api_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTasksRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Tasks",
        "tags": [
          "tasks"
        ]
      }
    },
...
    "/api/tasks/delete": {
      "post": {
        "description": "删除任务\n\n**同时删掉临时文件**（`cancel_many_async` 负责）。只删记录的话，\n下载目录里会留下一堆 `video_<task_id>.m4s`，而且再也没人认得它们属于谁",
        "operationId": "delete_tasks_api_tasks_delete_post",
        "parameters": [
          {
            "in": "query",
            "name": "completed",
            "required": false,
            "schema": {
              "default": false,
              "title": "Completed",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskIdsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Tasks",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/list": {
      "get": {
        "description": "列出任务，可排序\n\n与 `GET /api/tasks` 的区别：那个是给「刷新页面恢复现场」用的全量快照（带事件游标），\n这个是给列表视图用的，可以排序与限量",
        "operationId": "list_tasks_api_tasks_list_get",
        "parameters": [
          {
            "in": "query",
            "name": "completed",
            "required": false,
            "schema": {
              "default": false,
              "title": "Completed",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "default": "created_time",
              "title": "Sort By",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ascending",
            "required": false,
            "schema": {
              "default": true,
              "title": "Ascending",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tasks",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/pause": {
      "post": {
        "description": "暂停：让 aria2 停下来，并把任务状态置为已暂停",
        "operationId": "pause_tasks_api_tasks_pause_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskIdsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PauseResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pause Tasks",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/resume": {
      "post": {
        "description": "恢复",
        "operationId": "resume_tasks_api_tasks_resume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskIdsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PauseResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Tasks",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/retry": {
      "post": {
        "description": "重新下载\n\n先 `reset` 把进度与分片记录清干净再 `recreate` —— 只改状态不清记录的话，\n续传会接着上一次那份已经作废的分片表走",
        "operationId": "retry_tasks_api_tasks_retry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskIdsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Tasks",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "operationId": "get_task_api_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task",
        "tags": [
          "tasks"
        ]
      }
    }
  }
}